refactor: move shared mock data
